fix: organize code : upload to ipfs using pinata
