+++ packages/components/FilePreview/ImagesFullViewModal.tsx
@@ -7,9 +7,9 @@ import {
 } from "react-native";
 import { SvgProps } from "react-native-svg";
 
-import { ipfsURLToHTTPURL } from "./ipfs";
 import chevronLeft from "../../../assets/icons/chevron-left.svg";
 import chevronRight from "../../../assets/icons/chevron-right.svg";
+import { ipfsURLToHTTPURL } from "../../utils/ipfs";
 import { neutral22, neutral33 } from "../../utils/style/colors";
 import { SVG } from "../SVG";
 import ModalBase from "../modals/ModalBase";
