+++ packages/screens/Settings/SettingsScreen.tsx
@@ -25,7 +25,7 @@ import { neutralA3, primaryColor } from "../../utils/style/colors";
 import { fontSemibold14 } from "../../utils/style/fonts";
 
 const NFTAPIKeyInput: React.FC = () => {
-  const NFTApiKey = useSelector(selectNFTStorageAPI);
+  const userIPFSKey = useSelector(selectNFTStorageAPI);
   const dispatch = useAppDispatch();
   const commonStyles = useCommonStyles();
 
@@ -40,11 +40,12 @@ const NFTAPIKeyInput: React.FC = () => {
             },
           ]}
         >
-          NFT.Storage/Pinata.cloud API key (for Social Feed)
+          app.pinata.cloud JWT key (For file upload)
         </BrandText>
         <Pressable
           onPress={() =>
-            dispatch(setNFTStorageAPI(process.env.NFT_STORAGE_API || ""))
+            // We ask key at each upload for now (Don't have Teritori's key for now)
+            dispatch(setNFTStorageAPI(""))
           }
         >
           <Text style={[fontSemibold14, { color: primaryColor }]}>
@@ -55,7 +56,7 @@ const NFTAPIKeyInput: React.FC = () => {
       <SpacerColumn size={1.5} />
       <TextInput
         style={[commonStyles.apiInput]}
-        value={NFTApiKey}
+        value={userIPFSKey}
         onChangeText={(value) => dispatch(setNFTStorageAPI(value))}
       />
     </View>
